[native] Introduce PrestoToVeloxConnector
Allow dynamic registration of Presto to Velox connector adapters to enable support
for third party connectors. An additional adapter can be configured and
registered dynamically based on the configuration properties.
