+++ presto-native-execution/presto_cpp/main/PrestoServer.cpp
@@ -35,8 +35,8 @@
 #include "presto_cpp/main/operators/PartitionAndSerialize.h"
 #include "presto_cpp/main/operators/ShuffleRead.h"
 #include "presto_cpp/main/operators/UnsafeRowExchangeSource.h"
+#include "presto_cpp/main/types/PrestoToVeloxConnector.h"
 #include "presto_cpp/main/types/PrestoToVeloxQueryPlan.h"
-#include "presto_cpp/presto_protocol/ConnectorProtocol.h"
 #include "velox/common/base/Counters.h"
 #include "velox/common/base/StatsReporter.h"
 #include "velox/common/caching/CacheTTLController.h"
@@ -234,14 +234,14 @@ void PrestoServer::run() {
   registerShuffleInterfaceFactories();
   registerCustomOperators();
 
-  protocol::registerConnectorProtocol(
-      "hive", std::make_unique<protocol::HiveConnectorProtocol>());
-  protocol::registerConnectorProtocol(
-      "hive-hadoop2", std::make_unique<protocol::HiveConnectorProtocol>());
-  protocol::registerConnectorProtocol(
-      "iceberg", std::make_unique<protocol::IcebergConnectorProtocol>());
-  protocol::registerConnectorProtocol(
-      "tpch", std::make_unique<protocol::TpchConnectorProtocol>());
+  registerPrestoToVeloxConnector(
+      std::make_unique<HivePrestoToVeloxConnector>("hive"));
+  registerPrestoToVeloxConnector(
+      std::make_unique<HivePrestoToVeloxConnector>("hive-hadoop2"));
+  registerPrestoToVeloxConnector(
+      std::make_unique<IcebergPrestoToVeloxConnector>("iceberg"));
+  registerPrestoToVeloxConnector(
+      std::make_unique<TpchPrestoToVeloxConnector>("tpch"));
 
   initializeVeloxMemory();
   initializeThreadPools();
@@ -926,6 +926,9 @@ std::vector<std::string> PrestoServer::registerConnectors(
 
       PRESTO_STARTUP_LOG(INFO) << "Registering catalog " << catalogName
                                << " using connector " << connectorName;
+
+      // make sure connector type is supported
+      getPrestoToVeloxConnector(connectorName);
 
       std::shared_ptr<velox::connector::Connector> connector =
           facebook::velox::connector::getConnectorFactory(connectorName)
