+++ presto-native-execution/presto_cpp/main/types/CMakeLists.txt
@@ -14,8 +14,9 @@ add_library(presto_type_converter OBJECT TypeParser.cpp)
 
 target_link_libraries(presto_type_converter velox_type_parser)
 
-add_library(presto_types OBJECT PrestoToVeloxQueryPlan.cpp
-                                PrestoToVeloxExpr.cpp PrestoToVeloxSplit.cpp)
+add_library(
+  presto_types OBJECT PrestoToVeloxQueryPlan.cpp PrestoToVeloxExpr.cpp
+                      PrestoToVeloxSplit.cpp PrestoToVeloxConnector.cpp)
 add_dependencies(presto_types presto_operators presto_type_converter velox_type
                  velox_type_fbhive velox_dwio_dwrf_proto)
