+++ presto-native-execution/presto_cpp/main/types/tests/PlanConverterTest.cpp
@@ -21,8 +21,8 @@
 #include "presto_cpp/main/operators/PartitionAndSerialize.h"
 #include "presto_cpp/main/operators/ShuffleRead.h"
 #include "presto_cpp/main/operators/ShuffleWrite.h"
+#include "presto_cpp/main/types/PrestoToVeloxConnector.h"
 #include "presto_cpp/main/types/PrestoToVeloxQueryPlan.h"
-#include "presto_cpp/presto_protocol/ConnectorProtocol.h"
 #include "velox/connectors/hive/TableHandle.h"
 #include "velox/exec/tests/utils/TempDirectoryPath.h"
 
@@ -99,13 +99,23 @@ class PlanConverterTest : public ::testing::Test {
   static void SetUpTestCase() {
     memory::MemoryManager::testingSetInstance({});
   }
+
+  void SetUp() override {
+    registerPrestoToVeloxConnector(
+        std::make_unique<HivePrestoToVeloxConnector>("hive"));
+    registerPrestoToVeloxConnector(
+        std::make_unique<HivePrestoToVeloxConnector>("hive-plus"));
+  }
+
+  void TearDown() override {
+    unregisterPrestoToVeloxConnector("hive");
+    unregisterPrestoToVeloxConnector("hive-plus");
+  }
 };
 
 // Leaf stage plan for select regionkey, sum(1) from nation group by 1
 // Scan + Partial Agg + Repartitioning
 TEST_F(PlanConverterTest, scanAgg) {
-  protocol::registerConnectorProtocol(
-      "hive", std::make_unique<protocol::HiveConnectorProtocol>());
   auto partitionedOutput = assertToVeloxQueryPlan("ScanAgg.json");
   auto* tableScan = dynamic_cast<const core::TableScanNode*>(
       partitionedOutput->sources()[0]->sources()[0]->sources()[0].get());
@@ -132,8 +142,6 @@ TEST_F(PlanConverterTest, scanAgg) {
   ASSERT_EQ(tableParameters.find("totalSize")->second, "1451");
   ASSERT_EQ(tableParameters.find("foobar"), tableParameters.end());
 
-  protocol::registerConnectorProtocol(
-      "hive-plus", std::make_unique<protocol::HiveConnectorProtocol>());
   assertToVeloxQueryPlan("ScanAggCustomConnectorId.json");
 }
 
@@ -168,9 +176,6 @@ TEST_F(PlanConverterTest, offsetLimit) {
 }
 
 TEST_F(PlanConverterTest, batchPlanConversion) {
-  protocol::unregisterConnectorProtocol("hive");
-  protocol::registerConnectorProtocol(
-      "hive", std::make_unique<protocol::HiveConnectorProtocol>());
   filesystems::registerLocalFileSystem();
   auto root = assertToBatchVeloxQueryPlan(
       "ScanAggBatch.json",
